complate pricelist page
fix some bug on add_pro ,admin,layout_header

+++ admin.php
@@ -120,7 +120,7 @@ p.cat_id = c.cat_id and c.cat_id = 1 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -154,7 +154,7 @@ p.cat_id = c.cat_id and c.cat_id = 2 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -189,7 +189,7 @@ p.cat_id = c.cat_id and c.cat_id = 3 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -223,7 +223,7 @@ p.cat_id = c.cat_id and c.cat_id = 4 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -257,7 +257,7 @@ p.cat_id = c.cat_id and c.cat_id = 5 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -291,7 +291,7 @@ p.cat_id = c.cat_id and c.cat_id = 6 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -325,7 +325,7 @@ p.cat_id = c.cat_id and c.cat_id = 7 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -359,7 +359,7 @@ p.cat_id = c.cat_id and c.cat_id = 8 ORDER BY $sortby";
 						echo "<td>".$row['pro_pdis']."</td>";
 					}
 					echo "<td><a href='edit_pro.php?edit_id=".$row['pro_id']."'><img src='img/pro_edit.png' width='24' height='24'></td>";
-					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
+					echo "<td><a href='del_pro.php?delete_id=".$row['pro_id']."' class='confirmation'><img src='img/pro_delete.png' width='24' height='24'></a></td>";
 					echo "</tr>";
 				}
 				echo "</table>";
@@ -370,3 +370,8 @@ p.cat_id = c.cat_id and c.cat_id = 8 ORDER BY $sortby";
 <?php
 	echo $layout_footer->output();
 ?>
+<script type="text/javascript">
+    $('.confirmation').on('click', function () {
+        return confirm('Are you sure?');
+    });
+</script>
